LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/anchor-updates-after-explicit-scroll-expected.txt


FAIL Anchor node recomputed after an explicit scroll occurs. assert_equals: expected 600 but got 500
